requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
Appium-Python-Client==5.2.4
arrow==1.3.0
asgiproxy==0.2.0
asgiref==3.10.0
astroid==4.0.1
attrs==25.4.0
Authlib==1.6.5
axe-playwright-python==0.1.5
bandit==1.8.6
bcrypt==5.0.0
bidict==0.23.1
billiard==4.2.2
black==25.9.0
boolean.py==5.0
boto3==1.40.48
botocore==1.40.48
CacheControl==0.14.3
cachetools==6.2.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
cryptography==46.0.2
cyclonedx-bom==7.1.0
cyclonedx-python-lib==9.1.0
databricks-sdk==0.67.0
databricks-sqlalchemy==2.0.1
defusedxml==0.7.1
detect-secrets==1.5.0
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dparse==0.6.4
ecdsa==0.19.1
fastapi==0.118.2
filelock==3.20.0
fqdn==1.5.1
frozenlist==1.8.0
git-remote-s3==0.2.5
google-auth==2.41.1
googleapis-common-protos==1.70.0
granian==2.5.5
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isoduration==20.11.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
kombu==5.5.4
lark==1.3.0
license-expression==30.4.4
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.7.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
openai
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
outcome==1.3.0.post0
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pgvector==0.4.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
playwright==1.55.0
plotly==6.3.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyee==13.0.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.0
PyNaCl==1.6.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.14.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
reflex==0.8.14
reflex-ag-grid==0.0.11
reflex-enterprise==0.3.4.post2
reflex-global-hotkey==1.2.3
reflex-hosting-cli==0.1.56
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rq==2.6.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.0
s3transfer==0.14.0
safety==3.6.2
safety-schemas==0.0.16
selenium==4.36.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlmodel==0.0.27
starlette==0.48.0
stevedore==5.5.0
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.12.0
toml==0.10.2
tomlkit==0.13.3
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.19.2
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-python-dateutil==2.9.0.20251008
types-redis==4.6.0.20241004
types-requests==2.32.4.20250913
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
webcolors==24.11.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0